requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
spacy==3.7.2
python-dotenv==1.0.0
flask==3.0.0
werkzeug==3.0.1